thêm kho

+++ database/migrations/2021_05_07_153617_add_column_point_to_points.php
@@ -15,7 +15,7 @@ class AddColumnPointToPoints extends Migration
     {
         Schema::table('points', function (Blueprint $table) {
             //
-            $table->bigInteger("point");
+            $table->float('point',10,2);
         });
     }
