+++ database/migrations/2021_05_12_152436_add_column_total_point_to_transactions.php
@@ -15,7 +15,7 @@ class AddColumnTotalPointToTransactions extends Migration
     {
         Schema::table('transactions', function (Blueprint $table) {
             //
-            $table->bigInteger('point')->unsigned();
+            $table->float('point',10,2)->unsigned();
         });
     }
